User slice

+++ backend/src/main/java/com/rrss/backend/service/ProductService.java
@@ -2,8 +2,10 @@ package com.rrss.backend.service;
 
 import com.rrss.backend.dto.AddProductRequest;
 import com.rrss.backend.dto.ProductDto;
-import com.rrss.backend.model.Merchant;
+import com.rrss.backend.dto.UpdateProductRequest;
 import com.rrss.backend.model.Product;
+import com.rrss.backend.model.User;
+import com.rrss.backend.repository.ProductCategoryRepository;
 import com.rrss.backend.repository.ProductRepository;
 import com.rrss.backend.util.ImageUtil;
 import com.rrss.backend.util.UserUtil;
@@ -22,19 +24,23 @@ public class ProductService {
     private final ProductRepository repository;
     private final ProductCategoryService productCategoryService;
     private final UserUtil userUtil;
+    private final ProductCategoryRepository productCategoryRepository;
 
-    public ProductService(ProductRepository repository, ProductCategoryService productCategoryService, UserUtil userUtil) {
+    public ProductService(ProductRepository repository, ProductCategoryService productCategoryService, UserUtil userUtil, ProductCategoryRepository productCategoryRepository) {
         this.repository = repository;
         this.productCategoryService = productCategoryService;
         this.userUtil = userUtil;
+        this.productCategoryRepository = productCategoryRepository;
     }
 
     public ProductDto addProduct(Principal currentUser, AddProductRequest addProductRequest, MultipartFile file) throws IOException {
 
+        User user = userUtil.extractUser(currentUser);
+
         Product product = new Product(
                 addProductRequest.name(),
                 addProductRequest.description(),
-                Objects.requireNonNull(userUtil.extractUser(currentUser).getMerchant()),
+                user.getMerchant(),
                 productCategoryService.findByName(addProductRequest.productCategoryName()),
                 addProductRequest.price(),
                 ImageUtil.compressImage(file.getBytes())
@@ -43,13 +49,19 @@ public class ProductService {
         return ProductDto.convert(repository.save(product));
     }
 
-    public String deleteProduct(Principal currentUser, Long productId) {
+    public ProductDto deleteProduct(Principal currentUser, Long productId) {
         Product product = repository.findById(productId)
                 .orElseThrow(() -> new IllegalArgumentException("Product not found"));
 
+        User user = userUtil.extractUser(currentUser);
+
+        if (user.getMerchant() == null || Objects.equals(user.getMerchant().getId(), product.getMerchant().getId())) {
+            throw new IllegalArgumentException("you are not the owner of this product");
+        }
+
         repository.delete(product);
 
-        return "Product deleted successfully";
+        return ProductDto.convert(product);
     }
 
     public ProductDto getProduct(Long productId) {
@@ -84,5 +96,28 @@ public class ProductService {
                 .stream()
                 .map(ProductDto::convert)
                 .toList();
+    }
+
+    public ProductDto updateProduct(Principal currentUser, Long productId, UpdateProductRequest updateProductRequest) {
+        Product product = repository.findById(productId)
+                .orElseThrow(() -> new IllegalArgumentException("Product not found"));
+
+        User user = userUtil.extractUser(currentUser);
+        if (user.getMerchant() == null || Objects.equals(user.getMerchant().getId(), product.getMerchant().getId())) {
+            throw new IllegalArgumentException("you are not the owner of this product");
+        }
+
+                new Product(
+                productId,
+                updateProductRequest.name(),
+                updateProductRequest.description(),
+                product.getMerchant(),
+                productCategoryRepository.findByName(updateProductRequest.productCategoryName())
+                        .orElseThrow(() -> new IllegalArgumentException("Product category not found")),
+                updateProductRequest.price(),
+                product.getPicture()
+        );
+
+        return ProductDto.convert(repository.save(product));
     }
 }
